backend/requirements.txt
# Backend Dependencies for Todo Full-Stack Web Application

# FastAPI Framework
fastapi==0.110.0
uvicorn[standard]==0.27.1

# Database
sqlmodel==0.0.14
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.9

# Environment Configuration
python-dotenv==1.0.1
pydantic-settings==2.1.0

# Validation
email-validator>=2.0.0

# AI Agent
openai-agents>=0.0.7

# Testing
pytest==8.0.0
pytest-asyncio==0.23.5
httpx==0.26.0
